add-item-request fix again

+++ src/test/java/ru/practicum/shareit/item/controller/ItemControllerTest.java
@@ -38,8 +38,8 @@ public class ItemControllerTest {
     @Autowired
     private ObjectMapper objectMapper;
 
-    private static final Long USER_ID = 1;
-    private static final Long ITEM_ID = 2;
+    private static final Long USER_ID = 1L;
+    private static final Long ITEM_ID = 2L;
 
     @Test
     public void testGetItemById() throws Exception {
@@ -51,7 +51,7 @@ public class ItemControllerTest {
 
         mockMvc.perform(get("/items/" + ITEM_ID).header("X-Sharer-User-Id", USER_ID))
                 .andExpect(status().isOk())
-                .andExpect(jsonPath("$.id", is((int) ITEM_ID)))
+                .andExpect(jsonPath("$.id", is(ITEM_ID)))
                 .andExpect(jsonPath("$.name", is("Test item")));
     }
 
@@ -62,7 +62,7 @@ public class ItemControllerTest {
         item1.setName("Test item 1");
 
         ItemReplyDto item2 = new ItemReplyDto();
-        item2.setId(ITEM_ID + 1);
+        item2.setId(ITEM_ID + 1L);
         item2.setName("Test item 2");
 
         List<ItemReplyDto> items = Arrays.asList(item1, item2);
@@ -71,9 +71,9 @@ public class ItemControllerTest {
 
         mockMvc.perform(get("/items").header("X-Sharer-User-Id", USER_ID))
                 .andExpect(status().isOk())
-                .andExpect(jsonPath("$[0].id", is((int) ITEM_ID)))
+                .andExpect(jsonPath("$[0].id", is(ITEM_ID)))
                 .andExpect(jsonPath("$[0].name", is("Test item 1")))
-                .andExpect(jsonPath("$[1].id", is((int) ITEM_ID + 1)))
+                .andExpect(jsonPath("$[1].id", is(ITEM_ID + 1L)))
                 .andExpect(jsonPath("$[1].name", is("Test item 2")));
     }
 
@@ -84,7 +84,7 @@ public class ItemControllerTest {
         item1.setName("Test item 1");
 
         ItemReplyDto item2 = new ItemReplyDto();
-        item2.setId(ITEM_ID + 1);
+        item2.setId(ITEM_ID + 1L);
         item2.setName("Test item 2");
 
         List<ItemReplyDto> items = Arrays.asList(item1, item2);
@@ -93,10 +93,10 @@ public class ItemControllerTest {
 
         mockMvc.perform(get("/items/search").param("text", "test"))
                 .andExpect(status().isOk())
-                .andExpect(jsonPath("$[0].id", is((int) ITEM_ID)))
+                .andExpect(jsonPath("$[0].id", is((ITEM_ID)))
                 .andExpect(jsonPath("$[0].name", is("Test item 1")))
-                .andExpect(jsonPath("$[1].id", is((int) ITEM_ID + 1)))
-                .andExpect(jsonPath("$[1].name", is("Test item 2")));
+                .andExpect(jsonPath("$[1].id", is(ITEM_ID + 1L)))
+                .andExpect(jsonPath("$[1].name", is("Test item 2"))));
     }
 
     @Test
